voxel-ray intersect

+++ src/VX3/VX3_VoxelyzeKernel.cu
@@ -902,7 +902,7 @@ __global__ void gpu_update_occlusion(VX3_Voxel **surface_voxels, int num, VX3_Vo
             // todo: if the body spins around then this needs to be flipped
 
             // unit direction vector of ray
-            VX3_Vec3D<float> unitdir = k->LightPos - ray_origin ;  // ray_origin ---> k->LightPos
+            VX3_Vec3D<double> unitdir = k->LightPos - ray_origin ;  // ray_origin ---> k->LightPos
             unitdir = unitdir.Normalized();
 
             VX3_Vec3D<float> dirfrac;
